Move DifferentialIK class to experiment, and implement controller class in templates.py
